+++ packages/react-composites/src/composites/MeetingComposite/MeetingComposite.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import React, { useCallback, useState, useMemo } from 'react';
+import React, { useCallback, useState, useMemo, useEffect } from 'react';
 import { PartialTheme, Stack, Theme } from '@fluentui/react';
 import { CallComposite, CallControlOptions } from '../CallComposite';
 import { CallAdapterProvider } from '../CallComposite/adapter/CallAdapterProvider';
@@ -18,6 +18,7 @@ import { ChatCompositeProps } from '../ChatComposite';
 import { BaseComposite, BaseCompositeProps } from '../common/BaseComposite';
 import { CallCompositeIcons, ChatCompositeIcons } from '../common/icons';
 import { AvatarPersonaDataCallback } from '../common/AvatarPersona';
+import { ChatAdapterProvider } from '../ChatComposite/adapter/ChatAdapterProvider';
 
 /**
  * Props required for the {@link MeetingComposite}
@@ -67,10 +68,7 @@ export type MeetingCompositeOptions = {
  * @beta
  */
 export interface MeetingCallControlOptions
-  extends Pick<
-    CallControlOptions,
-    'cameraButton' | 'microphoneButton' | 'screenShareButton' | 'devicesButton' | 'displayType'
-  > {
+  extends Pick<CallControlOptions, 'cameraButton' | 'microphoneButton' | 'screenShareButton' | 'displayType'> {
   /**
    * Show or hide the chat button in the meeting control bar.
    * @defaultValue true
@@ -102,13 +100,19 @@ const MeetingScreen = (props: MeetingScreenProps): JSX.Element => {
 
   const [currentMeetingState, setCurrentMeetingState] = useState<CallState>();
   const [currentPage, setCurrentPage] = useState<MeetingCompositePage>();
-  meetingAdapter.onStateChange((newState) => {
-    setCurrentPage(newState.page);
-    setCurrentMeetingState(newState.meeting?.state);
-  });
-
   const [showChat, setShowChat] = useState(false);
   const [showPeople, setShowPeople] = useState(false);
+
+  useEffect(() => {
+    const updateMeetingPage = (newState): void => {
+      setCurrentPage(newState.page);
+      setCurrentMeetingState(newState.meeting?.state);
+    };
+    meetingAdapter.onStateChange(updateMeetingPage);
+    return () => {
+      meetingAdapter.offStateChange(updateMeetingPage);
+    };
+  }, [meetingAdapter]);
 
   const closePane = useCallback(() => {
     setShowChat(false);
@@ -170,16 +174,19 @@ const MeetingScreen = (props: MeetingScreenProps): JSX.Element => {
         )}
       </Stack>
       {(isInLobbyOrConnecting || hasJoinedCall) && (
-        <MeetingCallControlBar
-          callAdapter={callAdapter}
-          chatButtonChecked={showChat}
-          onChatButtonClicked={toggleChat}
-          peopleButtonChecked={showPeople}
-          onPeopleButtonClicked={togglePeople}
-          mobileView={props.formFactor === 'mobile'}
-          disableButtonsForLobbyPage={isInLobbyOrConnecting}
-          callControls={props.callControls}
-        />
+        <ChatAdapterProvider adapter={chatProps.adapter}>
+          <MeetingCallControlBar
+            callAdapter={callAdapter}
+            chatAdapter={chatProps.adapter}
+            chatButtonChecked={showChat}
+            onChatButtonClicked={toggleChat}
+            peopleButtonChecked={showPeople}
+            onPeopleButtonClicked={togglePeople}
+            mobileView={props.formFactor === 'mobile'}
+            disableButtonsForLobbyPage={isInLobbyOrConnecting}
+            callControls={props.callControls}
+          />
+        </ChatAdapterProvider>
       )}
     </Stack>
   );